requirements.txt
flask
flask-wtf
wtforms
flask-sqlalchemy
flask-migrate
requests
pillow